Skip migrations when Telescope is disabled

+++ database/migrations/2018_08_08_100000_create_telescope_entries_table.php
@@ -19,6 +19,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! config('telescope.enabled')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->create('telescope_entries', function (Blueprint $table) {
@@ -61,6 +65,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! config('telescope.enabled')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->dropIfExists('telescope_entries_tags');
